--- board-final-back/src/main/java/com/study/boardfinalback/error/BoardNotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class BoardNotFoundException extends BusinessException {
-    public BoardNotFoundException() {
-        super(Messages.NO_BOARD_MESSAGE);
-    }
-}
