--- board-final-back/src/main/java/com/study/boardfinalback/error/BusinessException.java
@@ -1,8 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class BusinessException extends RuntimeException {
-    protected BusinessException(String message) {
-        super(message);
-    }
-}
-
