--- board-final-back/src/main/java/com/study/boardfinalback/error/CategoryNotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class CategoryNotFoundException extends BusinessException {
-    public CategoryNotFoundException() {
-        super(Messages.NO_CATEGORY_MESSAGE);
-    }
-}
