--- board-final-back/src/main/java/com/study/boardfinalback/error/CommentNotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class CommentNotFoundException extends BusinessException {
-    public CommentNotFoundException() {
-        super(Messages.NO_COMMENT_MESSAGE);
-    }
-}
