--- board-final-back/src/main/java/com/study/boardfinalback/error/FileNotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class FileNotFoundException extends BusinessException {
-    public FileNotFoundException() {
-        super(Messages.NO_FILE_MESSAGE);
-    }
-}
