--- board-final-back/src/main/java/com/study/boardfinalback/error/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.study.boardfinalback.error;
-
-public class UserNotFoundException extends BusinessException {
-    public UserNotFoundException() {
-        super(Messages.NO_USER_MESSAGE);
-    }
-}
