+++ board-final-back/src/main/java/com/study/boardfinalback/error/comments/CommentNotFoundException.java
@@ -1,0 +1,10 @@
+package com.study.boardfinalback.error.comments;
+
+import com.study.boardfinalback.error.common.Messages;
+import com.study.boardfinalback.error.common.ResourceNotFoundException;
+
+public class CommentNotFoundException extends ResourceNotFoundException {
+    public CommentNotFoundException() {
+        super(Messages.NO_COMMENT_MESSAGE);
+    }
+}
