+++ board-final-back/src/main/java/com/study/boardfinalback/error/common/Messages.java
@@ -1,4 +1,4 @@
-package com.study.boardfinalback.error;
+package com.study.boardfinalback.error.common;
 
 public class Messages {
 
@@ -7,5 +7,9 @@ public class Messages {
     public static final String NO_CATEGORY_MESSAGE = "존재하지 않는 카테고리입니다.";
     public static final String NO_COMMENT_MESSAGE = "존재하지 않는 댓글입니다.";
     public static final String NO_FILE_MESSAGE = "존재하지 않는 파일입니다.";
+    public static final String DUPLICATED_USER_MESSAGE = "이미 존재하는 사용자입니다.";
+    public static final String UNAUTHENTICATED_USER_MESSAGE = "인증이 필요합니다.";
+    public static final String UNAUTHORIZED_USER_MESSAGE = "권한이 없습니다.";
+    public static final String PASSWORD_CHANGE_DTO_INVALID_MESSAGE = "올바르지 않은 비밀번호 형식입니다.";
 
 }
