+++ board-final-back/src/main/java/com/study/boardfinalback/error/common/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.study.boardfinalback.error.common;
+
+public class ResourceNotFoundException extends RuntimeException {
+    protected ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
+
