+++ board-final-back/src/main/java/com/study/boardfinalback/error/users/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package com.study.boardfinalback.error.users;
+
+import com.study.boardfinalback.error.common.Messages;
+import com.study.boardfinalback.error.common.ResourceNotFoundException;
+
+public class UserNotFoundException extends ResourceNotFoundException {
+    public UserNotFoundException() {
+        super(Messages.NO_USER_MESSAGE);
+    }
+}
