+++ board-final-back/src/main/java/com/study/boardfinalback/service/UserService.java
@@ -1,7 +1,7 @@
 package com.study.boardfinalback.service;
 
 import com.study.boardfinalback.domain.user.User;
-import com.study.boardfinalback.error.UserNotFoundException;
+import com.study.boardfinalback.error.users.UserNotFoundException;
 import com.study.boardfinalback.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
